Move ListQueryParameterUtils and test to infrastructure
