+++ infrastructure/restapi/src/test/java/tech/pegasys/teku/infrastructure/restapi/endpoints/ListQueryParameterUtilsTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.beaconrestapi;
+package tech.pegasys.teku.infrastructure.restapi.endpoints;
 
 import static org.assertj.core.api.Assertions.assertThat;
